Add Flyway / move rename files + update api url
