+++ src/main/resources/generator/server/springboot/dbmigration/flyway/resources/V00000000000000__init.sql
